requirements.txt
uvicorn
fastapi
sqlalchemy
psycopg2
SQLAlchemy-Utils
alembic
python-dotenv
pydantic[email]